Set up Icon
Set up Icon and made default crosshairs HUDItem
