+++ build.gradle
@@ -20,7 +20,7 @@ plugins {
     id "net.minecraftforge.gradle.forge" version "2.0.2"
 }
 */
-version = "1A1-3"
+version = "1A1-4"
 group= "net.binaryaura.customize" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "customize"
