+++ src/main/java/net/binaryaura/customize/client/gui/huditem/HudItemAdvancedText.java
@@ -7,17 +7,13 @@ public abstract class HudItemAdvancedText extends HudItem{
 		super(name);
 	}
 
-	protected abstract int getBGColor(int line);
-
-	protected abstract int getColor(int line);
-
-	protected abstract int getBGAlpha();
-
-	protected abstract int getAlpha();
-
 	@Override
 	protected void init() {
 
 	}
 
+	protected abstract int getBGAlpha();
+	protected abstract int getBGColor(int line);
+	protected abstract int getAlpha();
+	protected abstract int getColor(int line);
 }
