--- src/main/java/net/binaryaura/customize/client/gui/huditem/HudItemCrosshairs.java
@@ -1,19 +1,0 @@
-package net.binaryaura.customize.client.gui.huditem;
-
-public class HudItemCrosshairs extends HudItemIcon {
-
-	//	TODO: Make Crosshairs Icon HudItem
-	public HudItemCrosshairs(String name) {
-		super(name);
-	}
-
-	@Override
-	protected void init() {
-		
-	}
-
-	protected void setHeightAndWidth() {
-		
-	}
-
-}
