--- src/main/java/net/binaryaura/customize/client/gui/huditem/HudItemHotbar.java
@@ -1,22 +1,0 @@
-package net.binaryaura.customize.client.gui.huditem;
-
-public class HudItemHotbar extends HudItemIconSet {
-
-	//	TODO: Make Hotbar HudItem
-	//		Text for No. in stack
-	//		Icon for Cursor
-	//		IconGauge for Background and items
-	public HudItemHotbar(String name) {
-		super(name);
-	}
-
-	@Override
-	protected void init() {
-		
-	}
-
-	protected void setHeightAndWidth() {
-		
-	}
-
-}
