+++ src/main/java/net/binaryaura/customize/client/gui/huditem/HudItemIconGauge.java
@@ -3,50 +3,26 @@ package net.binaryaura.customize.client.gui.huditem;
 import net.binaryaura.customize.client.gui.LayeredSprite;
 import net.binaryaura.customize.client.gui.Sprite;
 import net.binaryaura.customize.client.gui.SpriteSet;
-import net.binaryaura.customize.client.gui.huditem.HudItemManager.HudItemType;
 import net.minecraft.util.MathHelper;
 
 public abstract class HudItemIconGauge extends HudItem {
 
+	protected static final int DFLT_DEMO_AMT = 20;
+	protected static final int DFLT_MAX_PER_ROW = 10;
+	protected static final int DFLT_MAX_STK_SPC = 11;
+	protected static final int DFLT_MIN_STK_SPC = 3;
+	protected static final int DFLT_SPC = 8;
+	
 	//	TODO: Fix Button when Rotating with IconGauge Stacks
 	//	TODO: Fix Border Collision with IconGauge
 	public HudItemIconGauge(String name) {
 		super(name);
-		type = HudItemType.ICON_GUAGE;
 		canFlip = true;
 		canRotate = true;
-	}
-	
-	public void setMaxPerRow(int max) {
-		maxPerRow = max;
-		width = (maxPerRow - 1)*space + layers.getWidth();
-		int stacks = MathHelper.ceiling_float_int(getAmount() / ((layers.getAmount() - 1) / maxPerRow));
-		int stackSpace = Math.max(maxStackSpace - (stacks - 1), minStackSpace);
-		height = (stacks - 1)*stackSpace + layers.getHeight();
-	}
-
-	@Override
-	protected void setHeightAndWidth() {
-					stacks = MathHelper.ceiling_float_int(amount / (layers.getAmount() - 1) / maxPerRow);
-					stackSpace = Math.max(maxStackSpace - (stacks - 1), minStackSpace);
-		switch(orientation) {
-			case RIGHT:
-					height = (stacks - 1)*stackSpace + layers.getHeight();
-					width = space*maxPerRow;
-				break;
-			case DOWN:
-					height = space*maxPerRow;
-					width = (stacks - 1)*stackSpace + layers.getWidth();
-				break;
-			case LEFT:
-					height = (stacks - 1)*stackSpace + layers.getHeight();
-					width = space*maxPerRow;
-				break;
-			case UP:
-					height = space*maxPerRow;
-					width = (stacks - 1)*stackSpace + layers.getWidth();
-				break;
-		}
+		maxPerRow = DFLT_MAX_PER_ROW;
+		maxStackSpace = DFLT_MAX_STK_SPC;
+		minStackSpace = DFLT_MIN_STK_SPC;
+		space = DFLT_SPC;
 	}
 
 	@Override
@@ -120,7 +96,15 @@ public abstract class HudItemIconGauge extends HudItem {
 		}
 		mc.mcProfiler.endSection();
 	}
-
+ 
+	public void setMaxPerRow(int max) {
+		maxPerRow = max;
+		width = (maxPerRow - 1)*space + layers.getWidth();
+		int stacks = MathHelper.ceiling_float_int(getAmount() / ((layers.getAmount() - 1) / maxPerRow));
+		int stackSpace = Math.max(maxStackSpace - (stacks - 1), minStackSpace);
+		height = (stacks - 1)*stackSpace + layers.getHeight();
+	}
+	
 	@Override
 	public int getButtonX() {
 		return super.getButtonX();
@@ -130,35 +114,33 @@ public abstract class HudItemIconGauge extends HudItem {
 	public int getButtonY() {
 		return super.getButtonY();
 	}
+	
+	@Override
+	protected void setHeightAndWidth() {
+					stacks = MathHelper.ceiling_float_int(amount / (layers.getAmount() - 1) / maxPerRow);
+					stackSpace = Math.max(maxStackSpace - (stacks - 1), minStackSpace);
+		switch(orientation) {
+			case RIGHT:
+					height = (stacks - 1)*stackSpace + layers.getHeight();
+					width = space*maxPerRow;
+				break;
+			case DOWN:
+					height = space*maxPerRow;
+					width = (stacks - 1)*stackSpace + layers.getWidth();
+				break;
+			case LEFT:
+					height = (stacks - 1)*stackSpace + layers.getHeight();
+					width = space*maxPerRow;
+				break;
+			case UP:
+					height = space*maxPerRow;
+					width = (stacks - 1)*stackSpace + layers.getWidth();
+				break;
+		}
+	}
 
-	protected abstract SpriteSet getIconSpriteSet(int icon);
-	
-	protected abstract SpriteSet getDefaultSpriteSet();
-	
-	protected abstract int getIconDeltaPerp(int icon);
-	
-	protected abstract int getIconDeltaPara(int icon);
-	
-	protected abstract float getAmount();
-	
-	protected abstract float getDefaultAmount();
-	
 	protected int bidirectionalShake() {
 		return rand.nextInt(3) - 1;
-	}
-	
-	protected int shake() {
-		return 2*rand.nextInt(2);
-	}
-	
-	// TODO: Fix IconGauge Animation Methods
-	protected int wave(int icon) {
-		if(updateCounter % MathHelper.ceiling_float_int(getAmount() + 5) == icon) {
-			animationFinished = false;
-			if(MathHelper.ceiling_float_int(getAmount()) - 1 == icon) animationFinished = true;
-			return 2;
-		}
-		return 0;
 	}
 	
 	protected int movingHalfSinWave(int icon) {
@@ -169,20 +151,45 @@ public abstract class HudItemIconGauge extends HudItem {
 		return 0;
 	}
 	
+	protected int shake() {
+		return 2*rand.nextInt(2);
+	}
+	
 	protected int sinWave(int icon) {
 		int leadIcon = updateCounter % MathHelper.ceiling_float_int(getAmount());
 		return MathHelper.ceiling_double_int(5*Math.sin(leadIcon*5 / Math.PI));
 	}
+	
+	// TODO: Fix IconGauge Animation Methods
+	protected int wave(int icon) {
+		if(updateCounter % MathHelper.ceiling_float_int(getAmount() + 5) == icon) {
+			animationFinished = false;
+			if(MathHelper.ceiling_float_int(getAmount()) - 1 == icon)
+				animationFinished = true;
+			return 2;
+		}
+		return 0;
+	}
+	
+	protected abstract int getIconDeltaPara(int icon);	
+	protected abstract int getIconDeltaPerp(int icon);	
+	protected abstract float getAmount();	
+	protected abstract float getDemoAmount();	
+	protected abstract SpriteSet getDemoSpriteSet();
+	protected abstract SpriteSet getIconSpriteSet(int icon);
+	
+	@Deprecated
+	protected boolean animationFinished = true;
+	
+	// Defaults Based on the default vitals bars
+	protected int maxPerRow;
+	protected int maxStackSpace;
+	protected int minStackSpace;
+	protected int space;
+	
+	protected LayeredSprite layers;
 
 	private int stacks = 0;
 	private int stackSpace = 0;
 	private float amount = 0;
-	protected boolean animationFinished = true;
-	
-	// Defaults Based on the default vitals bars
-	protected int maxStackSpace = 11;
-	protected int minStackSpace = 3;
-	protected int space = 8;
-	protected int maxPerRow = 10;
-	protected LayeredSprite layers;
 }
