--- src/main/java/net/binaryaura/customize/client/gui/huditem/HudItemPlayerList.java
@@ -1,52 +1,0 @@
-package net.binaryaura.customize.client.gui.huditem;
-
-public class HudItemPlayerList extends HudItemText {
-
-	//	TODO: Make Player List HudItem
-	//		Icon/IconList for player heads
-	//		IconGauge for Signal
-	//		Text for names
-	public HudItemPlayerList(String name) {
-		super(name);
-	}
-
-	@Override
-	protected void init() {
-		
-	}
-
-	@Override
-	protected int getDeltaX() {
-		return 0;
-	}
-
-	@Override
-	protected int getDeltaY() {
-		return 0;
-	}
-
-	@Override
-	protected int getBGColor() {
-		return 0;
-	}
-
-	@Override
-	protected int getColor() {
-		return 0;
-	}
-
-	@Override
-	protected int getBGAlpha() {
-		return 0;
-	}
-
-	@Override
-	protected int getAlpha() {
-		return 0;
-	}
-
-	protected void setHeightAndWidth() {
-		
-	}
-
-}
