+++ src/main/java/net/binaryaura/customize/client/gui/huditem/defaults/HudItemMountHealth.java
@@ -1,8 +1,15 @@
-package net.binaryaura.customize.client.gui.huditem;
+package net.binaryaura.customize.client.gui.huditem.defaults;
 
 import net.binaryaura.customize.client.gui.SpriteSet;
+import net.binaryaura.customize.client.gui.huditem.HudItemIconGauge;
 
 public class HudItemMountHealth extends HudItemIconGauge {
+	
+	private static final int DFLT_DEMO_AMT = 20;
+	private static final int DFLT_X = 0;
+	private static final int DFLT_Y = 0;
+	private static final Anchor DFLT_ANCH = Anchor.TOPLEFT;
+	private static final Orientation DFLT_ORIEN = Orientation.RIGHT;
 
 	//	TODO: Make Mount Health Gauge HudItem
 	public HudItemMountHealth(String name) {
@@ -15,7 +22,7 @@ public class HudItemMountHealth extends HudItemIconGauge {
 	}
 
 	@Override
-	protected SpriteSet getDefaultSpriteSet() {
+	protected SpriteSet getDemoSpriteSet() {
 		return null;
 	}
 
@@ -40,7 +47,7 @@ public class HudItemMountHealth extends HudItemIconGauge {
 	}
 
 	@Override
-	protected float getDefaultAmount() {
-		return 0;
+	protected float getDemoAmount() {
+		return DFLT_DEMO_AMT;
 	}
 }
