+++ src/main/java/net/binaryaura/customize/client/gui/huditem/defaults/HudItemSubtitle.java
@@ -1,7 +1,13 @@
-package net.binaryaura.customize.client.gui.huditem;
+package net.binaryaura.customize.client.gui.huditem.defaults;
+
+import net.binaryaura.customize.client.gui.huditem.HudItemText;
 
 public class HudItemSubtitle extends HudItemText {
 
+	private static final int DFLT_X = 0;
+	private static final int DFLT_Y = 0;
+	private static final Anchor DFLT_ANCH = Anchor.TOPLEFT;	
+	
 	//	TODO: Make Subtitle HudItem
 	public HudItemSubtitle(String name) {
 		super(name);
