+++ src/main/java/net/binaryaura/customize/common/Customize.java
@@ -19,7 +19,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 public class Customize {
     public static final String MODID = "customize";
 	public static final String NAME = "Customize";
-    public static final String VERSION = "1A1-3";
+    public static final String VERSION = "1A1-4";
     public static final String CLIENTPROXY = "net.binaryaura.customize.client.";
     public static final String COMMONPROXY = "net.binaryaura.customize.common.";
     
@@ -52,6 +52,7 @@ public class Customize {
     @EventHandler
     public void disable(FMLModDisabledEvent event) {
     	mc.ingameGUI = new GuiIngameForge(mc);
+    	log.info("Successfully unloaded " + NAME);
     }
     
 }
